fastapi_api/requirements.txt
arrow==1.2.0
alembic==1.10.2
Deprecated==1.2.13
fastapi==0.108.0
python-dotenv==0.19.2
uvicorn==0.22.0
pydantic==2.0.3
starlette==0.32.0
starlette-context==0.3.6
sqlalchemy-mixins==2.0.5
psycopg2-binary
pyjwt
bcrypt